+++ src/kodo_rlnc_c/kodo_rlnc_c.cpp
@@ -14,288 +14,287 @@
 #include <kodo_rlnc/on_the_fly_codes.hpp>
 #include "on_the_fly_encoder_stack.hpp"
 #include "on_the_fly_decoder_stack.hpp"
-
-struct kodo_rlnc_decoder
-{
-    kodo_rlnc_c::on_the_fly_decoder_stack m_decoder;
-};
-
-struct kodo_rlnc_decoder_factory
-{
-    template<class... Args>
-    kodo_rlnc_decoder_factory(Args&&... args) :
-        m_decoder_factory(std::forward<Args>(args)...)
-    { }
-    kodo_rlnc_c::on_the_fly_decoder_stack::config m_decoder_factory;
-};
-
-struct kodo_rlnc_encoder
-{
-    kodo_rlnc_c::on_the_fly_encoder_stack m_encoder;
-};
-
-struct kodo_rlnc_encoder_factory
-{
-    template<class... Args>
-    kodo_rlnc_encoder_factory(Args&&... args) :
-        m_encoder_factory(std::forward<Args>(args)...)
-    { }
-
-    kodo_rlnc_c::on_the_fly_encoder_stack::config m_encoder_factory;
-};
-
+namespace {
 fifi::api::field build_finite_field(int32_t finite_field_id)
 {
     switch (finite_field_id)
     {
-        case kodo_rlnc_binary:
+        case krlnc_binary:
             return fifi::api::field::binary;
-        case kodo_rlnc_binary4:
+        case krlnc_binary4:
             return fifi::api::field::binary4;
-        case kodo_rlnc_binary8:
+        case krlnc_binary8:
             return fifi::api::field::binary8;
     default:
         assert(false && "Unknown field");
         return fifi::api::field::binary;
     }
 }
+}
+
+struct krlnc_decoder
+{
+    kodo_rlnc_c::on_the_fly_decoder_stack m_impl;
+};
+
+struct krlnc_decoder_factory
+{
+    template<class... Args>
+    krlnc_decoder_factory(Args&&... args) :
+        m_impl(std::forward<Args>(args)...)
+    { }
+    kodo_rlnc_c::on_the_fly_decoder_stack::config m_impl;
+};
+
+struct krlnc_encoder
+{
+    kodo_rlnc_c::on_the_fly_encoder_stack m_impl;
+};
+
+struct krlnc_encoder_factory
+{
+    template<class... Args>
+    krlnc_encoder_factory(Args&&... args) :
+        m_impl(std::forward<Args>(args)...)
+    { }
+
+    kodo_rlnc_c::on_the_fly_encoder_stack::config m_impl;
+};
 
 //------------------------------------------------------------------
 // ENCODER FACTORY API
 //------------------------------------------------------------------
 
-kodo_rlnc_encoder_factory_t kodo_rlnc_encoder_factory_construct(
+krlnc_encoder_factory_t krlnc_new_encoder_factory(
     int32_t finite_field_id, uint32_t symbols, uint32_t symbol_size)
 {
     auto finite_field = build_finite_field(finite_field_id);
-    return new kodo_rlnc_encoder_factory(finite_field, symbols, symbol_size);
+    return new krlnc_encoder_factory(finite_field, symbols, symbol_size);
 }
 
-void kodo_rlnc_encoder_factory_destruct(kodo_rlnc_encoder_factory_t factory)
+void krlnc_delete_encoder_factory(krlnc_encoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    delete api;
+    assert(factory != nullptr);
+    delete factory;
 }
 
-uint32_t kodo_rlnc_encoder_factory_symbols(kodo_rlnc_encoder_factory_t factory)
+uint32_t krlnc_encoder_factory_symbols(krlnc_encoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    return api->m_encoder_factory.symbols();
+    assert(factory != nullptr);
+    return factory->m_impl.symbols();
 }
 
-uint32_t kodo_rlnc_encoder_factory_symbol_size(kodo_rlnc_encoder_factory_t factory)
+uint32_t krlnc_encoder_factory_symbol_size(krlnc_encoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    return api->m_encoder_factory.symbol_size();
+    assert(factory != nullptr);
+    return factory->m_impl.symbol_size();
 }
 
-void kodo_rlnc_encoder_factory_set_symbols(kodo_rlnc_encoder_factory_t factory, uint32_t symbols)
+void krlnc_encoder_factory_set_symbols(krlnc_encoder_factory_t factory, uint32_t symbols)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    api->m_encoder_factory.set_symbols(symbols);
+    assert(factory != nullptr);
+    factory->m_impl.set_symbols(symbols);
 }
 
-void kodo_rlnc_encoder_factory_set_symbol_size(
-    kodo_rlnc_encoder_factory_t factory, uint32_t symbol_size)
+void krlnc_encoder_factory_set_symbol_size(
+    krlnc_encoder_factory_t factory, uint32_t symbol_size)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    api->m_encoder_factory.set_symbol_size(symbol_size);
+    assert(factory != nullptr);
+    factory->m_impl.set_symbol_size(symbol_size);
 }
 
-kodo_rlnc_encoder_t kodo_rlnc_encoder_factory_build(kodo_rlnc_encoder_factory_t factory)
+krlnc_encoder_t krlnc_encoder_factory_build(krlnc_encoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_encoder_factory*) factory;
-    auto encoder = new kodo_rlnc_encoder();
-    encoder->m_encoder.initialize(api->m_encoder_factory);
+    assert(factory != nullptr);
+    auto encoder = new krlnc_encoder();
+    encoder->m_impl.initialize(factory->m_impl);
     return encoder;
 }
 
-void kodo_rlnc_encoder_destruct(kodo_rlnc_encoder_t encoder)
+void krlnc_delete_encoder(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    delete api;
+    assert(encoder != nullptr);
+    delete encoder;
 }
 
 //------------------------------------------------------------------
 // DECODER FACTORY API
 //------------------------------------------------------------------
 
-kodo_rlnc_decoder_factory_t kodo_rlnc_decoder_factory_construct(
+krlnc_decoder_factory_t krlnc_new_decoder_factory(
     int32_t finite_field_id, uint32_t symbols, uint32_t symbol_size)
 {
     auto finite_field = build_finite_field(finite_field_id);
-    return new kodo_rlnc_decoder_factory(finite_field, symbols, symbol_size);
+    return new krlnc_decoder_factory(finite_field, symbols, symbol_size);
 }
 
-void kodo_rlnc_decoder_factory_destruct(kodo_rlnc_decoder_factory_t factory)
+void krlnc_delete_decoder_factory(krlnc_decoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    delete api;
+    assert(factory != nullptr);
+    delete factory;
 }
 
-uint32_t kodo_rlnc_decoder_factory_symbols(kodo_rlnc_decoder_factory_t factory)
+uint32_t krlnc_decoder_factory_symbols(krlnc_decoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    return api->m_decoder_factory.symbols();
+    assert(factory != nullptr);
+    return factory->m_impl.symbols();
 }
 
-uint32_t kodo_rlnc_decoder_factory_symbol_size(kodo_rlnc_decoder_factory_t factory)
+uint32_t krlnc_decoder_factory_symbol_size(krlnc_decoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    return api->m_decoder_factory.symbol_size();
+    assert(factory != nullptr);
+    return factory->m_impl.symbol_size();
 }
 
-void kodo_rlnc_decoder_factory_set_symbols(kodo_rlnc_decoder_factory_t factory, uint32_t symbols)
+void krlnc_decoder_factory_set_symbols(krlnc_decoder_factory_t factory, uint32_t symbols)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    api->m_decoder_factory.set_symbols(symbols);
+    assert(factory != nullptr);
+    factory->m_impl.set_symbols(symbols);
 }
 
-void kodo_rlnc_decoder_factory_set_symbol_size(
-    kodo_rlnc_decoder_factory_t factory, uint32_t symbol_size)
+void krlnc_decoder_factory_set_symbol_size(
+    krlnc_decoder_factory_t factory, uint32_t symbol_size)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    api->m_decoder_factory.set_symbol_size(symbol_size);
+    assert(factory != nullptr);
+    factory->m_impl.set_symbol_size(symbol_size);
 }
 
-kodo_rlnc_decoder_t kodo_rlnc_decoder_factory_build(kodo_rlnc_decoder_factory_t factory)
+krlnc_decoder_t krlnc_decoder_factory_build(krlnc_decoder_factory_t factory)
 {
-    auto api = (kodo_rlnc_decoder_factory*) factory;
-    auto decoder = new kodo_rlnc_decoder();
-    decoder->m_decoder.initialize(api->m_decoder_factory);
+    assert(factory != nullptr);
+    auto decoder = new krlnc_decoder();
+    decoder->m_impl.initialize(factory->m_impl);
     return decoder;
 }
 
-void kodo_rlnc_decoder_destruct(kodo_rlnc_decoder_t decoder)
+void krlnc_delete_decoder(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    delete api;
+    assert(decoder != nullptr);
+    delete decoder;
 }
 
 //------------------------------------------------------------------
 // PAYLOAD API DECODER
 //------------------------------------------------------------------
 
-uint32_t kodo_rlnc_decoder_payload_size(kodo_rlnc_decoder_t decoder)
+uint32_t krlnc_decoder_payload_size(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.payload_size();
+    assert(decoder != nullptr);
+    return decoder->m_impl.payload_size();
 }
 
-void kodo_rlnc_decoder_read_payload(kodo_rlnc_decoder_t decoder, uint8_t* payload)
+void krlnc_decoder_read_payload(krlnc_decoder_t decoder, uint8_t* payload)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    api->m_decoder.read_payload(payload);
+    assert(decoder != nullptr);
+    decoder->m_impl.read_payload(payload);
 }
 
 //------------------------------------------------------------------
 // PAYLOAD API ENCODER
 //------------------------------------------------------------------
 
-uint32_t kodo_rlnc_encoder_payload_size(kodo_rlnc_encoder_t encoder)
+uint32_t krlnc_encoder_payload_size(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.payload_size();
+    assert(encoder != nullptr);
+    return encoder->m_impl.payload_size();
 }
 
-uint32_t kodo_rlnc_encoder_write_payload(kodo_rlnc_encoder_t encoder, uint8_t* payload)
+uint32_t krlnc_encoder_write_payload(krlnc_encoder_t encoder, uint8_t* payload)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.write_payload(payload);
+    assert(encoder != nullptr);
+    return encoder->m_impl.write_payload(payload);
 }
 
 //------------------------------------------------------------------
 // SYMBOL STORAGE API DECODER
 //------------------------------------------------------------------
 
-uint32_t kodo_rlnc_decoder_block_size(kodo_rlnc_decoder_t decoder)
+uint32_t krlnc_decoder_block_size(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.block_size();
+    assert(decoder != nullptr);
+    return decoder->m_impl.block_size();
 }
 
-uint32_t kodo_rlnc_decoder_symbol_size(kodo_rlnc_decoder_t decoder)
+uint32_t krlnc_decoder_symbol_size(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.symbol_size();
+    assert(decoder != nullptr);
+    return decoder->m_impl.symbol_size();
 }
 
-uint32_t kodo_rlnc_decoder_symbols(kodo_rlnc_decoder_t decoder)
+uint32_t krlnc_decoder_symbols(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.symbols();
+    assert(decoder != nullptr);
+    return decoder->m_impl.symbols();
 }
 
-void kodo_rlnc_decoder_set_mutable_symbols(
-    kodo_rlnc_decoder_t decoder, uint8_t* data, uint32_t size)
+void krlnc_decoder_set_mutable_symbols(
+    krlnc_decoder_t decoder, uint8_t* data, uint32_t size)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    api->m_decoder.set_mutable_symbols(storage::storage(data, size));
+    assert(decoder != nullptr);
+    decoder->m_impl.set_mutable_symbols(storage::storage(data, size));
 }
 
 //------------------------------------------------------------------
 // SYMBOL STORAGE API ENCODER
 //------------------------------------------------------------------
 
-uint32_t kodo_rlnc_encoder_block_size(kodo_rlnc_encoder_t encoder)
+uint32_t krlnc_encoder_block_size(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.block_size();
+    assert(encoder != nullptr);
+    return encoder->m_impl.block_size();
 }
 
-uint32_t kodo_rlnc_encoder_symbol_size(kodo_rlnc_encoder_t encoder)
+uint32_t krlnc_encoder_symbol_size(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.symbol_size();
+    assert(encoder != nullptr);
+    return encoder->m_impl.symbol_size();
 }
 
-uint32_t kodo_rlnc_encoder_symbols(kodo_rlnc_encoder_t encoder)
+uint32_t krlnc_encoder_symbols(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.symbols();
+    assert(encoder != nullptr);
+    return encoder->m_impl.symbols();
 }
 
-void kodo_rlnc_encoder_set_const_symbols(
-    kodo_rlnc_encoder_t encoder, uint8_t* data, uint32_t size)
+void krlnc_encoder_set_const_symbols(
+    krlnc_encoder_t encoder, uint8_t* data, uint32_t size)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    api->m_encoder.set_const_symbols(storage::storage(data, size));
+    assert(encoder != nullptr);
+    encoder->m_impl.set_const_symbols(storage::storage(data, size));
 }
 
 //------------------------------------------------------------------
 // DECODER API
 //------------------------------------------------------------------
 
-uint8_t kodo_rlnc_decoder_is_complete(kodo_rlnc_decoder_t decoder)
+uint8_t krlnc_decoder_is_complete(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.is_complete();
+    assert(decoder != nullptr);
+    return decoder->m_impl.is_complete();
 }
 
-uint32_t kodo_rlnc_decoder_rank(kodo_rlnc_decoder_t decoder)
+uint32_t krlnc_decoder_rank(krlnc_decoder_t decoder)
 {
-    auto api = (kodo_rlnc_decoder*) decoder;
-    return api->m_decoder.rank();
+    assert(decoder != nullptr);
+    return decoder->m_impl.rank();
 }
 
 //------------------------------------------------------------------
 // ENCODER API
 //------------------------------------------------------------------
 
-uint8_t kodo_rlnc_is_systematic_on(kodo_rlnc_encoder_t encoder)
+uint8_t krlnc_is_systematic_on(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    return api->m_encoder.is_systematic_on();
+    assert(encoder != nullptr);
+    return encoder->m_impl.is_systematic_on();
 }
 
-void kodo_rlnc_set_systematic_on(kodo_rlnc_encoder_t encoder)
+void krlnc_set_systematic_on(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    api->m_encoder.set_systematic_on();
+    encoder->m_impl.set_systematic_on();
 }
 
-void kodo_rlnc_encoder_set_systematic_off(kodo_rlnc_encoder_t encoder)
+void krlnc_encoder_set_systematic_off(krlnc_encoder_t encoder)
 {
-    auto api = (kodo_rlnc_encoder*) encoder;
-    api->m_encoder.set_systematic_off();
+    encoder->m_impl.set_systematic_off();
 }