+++ NEWS.rst
@@ -6,7 +6,9 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Upgraded to kodo-rlnc 11.
+* Major: Changed c "namespace" from ``kodo_rlnc`` to ``krlnc``.
+* Major: Changed ``construct`` and ``destruct`` to ``new`` and ``delete``.
 
 1.0.0
 -----